requirements.txt
flask
qrcode[pil]
python-dotenv
requests